[#101] refactor: 일부 리렌더링 리팩토링/예약 버튼 UI수정
